# TSX is not dead yet