… rpcs3/Emu/Cell/SPUThread.cpp
… rpcs3/Emu/Memory/vm_locking.h
… rpcs3/Emu/Memory/vm_reservation.h
… rpcs3/Emu/System.h
… rpcs3/Emu/system_config.h