I have rewritten logic for transaction fallback. Now it's based on wasted time, not some weird flags or useless counters. Or useless flags or weird counters? Doesn't matter.