@Aphelion42069 2020-10-31
This fixes the freezing during loading screens in Fallout 3 with TSX-FA (once again!)
![unknown](https://user-images.githubusercontent.com/61633628/97768919-9f300980-1af4-11eb-9c76-e39d86dc0e8c.png)

Spurs test has also improved with this PR - SPU SpinLock no longer takes ~5 minutes

Master:
SPU SpinLock completed in 249508 ms (PS3: 4409 ms)

PR:
SPU SpinLock completed in 12159 ms (PS3: 4409 ms)

Otherwise, performance seems to be the same as master in the few games I've tested